2_fastapi_postgres_react_jwt/backend/requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]